Logistic map and cobweb

+++ Coupled Logistic Maps/Graphing.cpp
@@ -107,7 +107,7 @@ FrameContext* WaitForNextFrameContext();
 LRESULT WINAPI WndProc(HWND hWnd, UINT msg, WPARAM wParam, LPARAM lParam);
 
 // Main code
-int Graph(void (*func)(float*, float*, int), float* mus, float* xs, int number)
+int Graph(void (*func)(float*, float*, float), float* mus, float* xs, float number)
 {
     // Make process DPI aware and obtain main monitor scale
     ImGui_ImplWin32_EnableDpiAwareness();
