+++ Coupled Logistic Maps/Graphing.h
@@ -1,3 +1,3 @@
 #pragma once
 
-int Graph(void (*func)(float*, float*, int), float*, float*, int);
+int Graph(void (*func)(float*, float*, float), float*, float*, float);
